Subir proyecto

+++ clase2/src/main/java/com/cibertec/clase2/repositories/AlumnoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlumnoRepository extends CrudRepository<Alumno, Long> {
+public interface AlumnoRepository extends CrudRepository<Alumno, Integer> {
 }
